Fix protocol ID mapping bugs and enable PeerSharing
- LocalTxMonitor: fix protocol ID from 12 to 9 (matches pallas/Haskell spec)
- MiniProtocolId enum: fix TxSubmission2 from 6→4, LocalTxSubmission from 9→6, LocalTxMonitor from 12→9
- TxSubmission2: fix ack_count/req_count parsing from u32 to u16 per spec
- Enable PeerSharing in N2N handshake (was disabled despite having server handler)
- Update networking docs with correct protocol ID table

+++ crates/torsten-network/src/n2n_server.rs
@@ -1224,8 +1224,8 @@ fn handle_n2n_txsubmission(
         // MsgRequestTxIds: [0, blocking, ack_count, req_count]
         0 => {
             let _blocking = decoder.bool().unwrap_or(false);
-            let ack_count = decoder.u32().unwrap_or(0) as usize;
-            let req_count = decoder.u32().unwrap_or(0) as usize;
+            let ack_count = decoder.u16().unwrap_or(0) as usize;
+            let req_count = decoder.u16().unwrap_or(0) as usize;
 
             // Acknowledge previously sent tx IDs (remove from inflight)
             if ack_count > 0 && ack_count <= peer_state.tx_inflight.len() {
